Improve Diviner performance with caching; relieve memory pressure
Summary:
Diviner reparses everything right now, which takes a minute or so. Use caching
to reduce this to a few seconds.

Ship files to engines in smaller batches (32 at once instead of all of them) to
relieve memory pressure. Use dispose() on XHPAST trees in the XHP engine.

Test Plan:
Generated Phabricator docs in a couple of seconds instead of a minute or so,
with peak memory usage of 25MB.

Reviewed By: jungejason
Reviewers: aran, tuomaspelkonen, jungejason
CC: aran, jungejason
Differential Revision: 206